Create README.md
INTRODUCTION:
To begin with, you need to know the importance of using a triad for a project of this nature. It is logical that the absence of one of the pillars will not prevent the achievement of pre-established objectives, however the concept involves its own guidelines for its adequacy. It is worth mentioning that the pillars can also be made more flexible, based on the legislation of each nation and its sources of direct law, such as customary law. Finally, I hope that many will benefit and that others will begin to comply with what each order requires, or even driven by issues of ethical conduct, make use of everything that will be explained.